% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_configuration.R
\name{checking_configuration}
\alias{checking_configuration}
\title{checking_configuration}
\usage{
checking_configuration(config_folder)
}
\arguments{
\item{config_foler}{The system based absolute path of this script (because
the configuration file needs to be next to it).}
}
\value{
String: the absolute path of the configuration file.
}
\description{
Checking the configuration file: if it exists, it does nothing.
If it doesn't, it throws an error.
}
\details{
Error codes:

ERROR CODE 1: The configuration file does not exist.
}
