% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_log_file.R
\name{checking_log_file}
\alias{checking_log_file}
\title{checking_log_file}
\usage{
checking_log_file(config_folder)
}
\arguments{
\item{config_folder}{The system-based absolute path of this script (because the configuration file needs to be
next to it.)}
}
\value{
The full log file path
}
\description{
Checking the log file and if it exists, it will delete it.
}
