% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log_file.R
\name{get_log_file}
\alias{get_log_file}
\title{checking_log_file}
\usage{
get_log_file(config_folder)
}
\arguments{
\item{config_folder}{The system-based absolute path of this script (because the configuration file needs to be next to it.)}
}
\value{
The full log file path
}
\description{
Create a timestamped log filename.
Example: preprocessing_2025-11-18_14-26-30.log
}
