% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_config_file.R
\name{loading.config.file}
\alias{loading.config.file}
\title{loading_config_file}
\usage{
loading.config.file(configuration.file.path, log.file)
}
\arguments{
\item{configuration_file_path}{The absolute path of the configuration file.}

\item{log_file}{The path of the log file.

Error codes:

ERROR CODE 2: Unable to read configuration file.
ERROR CODE 3: Configuration file is empty.}
}
\value{
Yaml object: contains the configuration file's content.
}
\description{
Loading the content from the configuration file.
}
