% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_log_message.R
\name{log}
\alias{log}
\title{add_log_message}
\usage{
log(message, log.file)
}
\arguments{
\item{message}{The actual commit message before push}

\item{log.file}{The path of the log file}
}
\description{
Log a message with a time stamp.
}
\keyword{internal}
