% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_main.R
\name{main}
\alias{main}
\title{main}
\usage{
main()
}
\description{
Main function for the script.
}
\details{
ERROR CODE 4: Invalid or missing 'input_dir' in configuration file
ERROR CODE 9: Missing 'Fastq_file_format' in configuration file.
ERROR CODE 10: Missing required genome preparation parameters in configuration file
ERROR CODE 12: Platform is '{platform}'. No processing available for this platform. Exiting
}
