% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_all_samples.R
\name{normalize_all_samples}
\alias{normalize_all_samples}
\title{normalize_all_samples}
\usage{
normalize_all_samples(output_dir, HVG_selection, number_top_genes, log_file)
}
\arguments{
\item{output_dir}{(str) Path to the main output directory containing per-sample subfolders.}

\item{log_file}{(str): Path to the central log file.}
}
\description{
Normalize all QC-filtered .h5ad files from each sample directory.
For each sample in the output_dir:
\itemize{
\item Loads its 'QC_filtered.rds' file
\item Performs total-count normalization (counts per 10,000)
\item Applies log1p transformation
\item Saves the normalized file as 'normalized.rds' in the same folder
}
}
\details{
ERROR CODE 13: Normalization failed.
}
