% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_start_genome.R
\name{prepare_STAR_genome}
\alias{prepare_STAR_genome}
\title{prepare_STAR_genome
Preparing the STAR genome index directory using full genome indexing and splice junction parameters.
Checks whether a valid STAR genome index already exists in the specified
genome directory. If all required index files are present, it skips index generation.
Otherwise, it generates a new STAR genome index using the provided FASTA, genome indexing
parameters, and splice junction parameters.}
\usage{
prepare_STAR_genome(
  genome_dir,
  fasta_file,
  genome_index_params,
  splice_junction_params,
  log_file
)
}
\arguments{
\item{genome_dir}{(str): The system-based absolute path of the directory where the STAR
genome index will be stored.}

\item{fasta_file}{(str): Path to the reference genome FASTA file.}

\item{genome_index_params}{(named list) Genome indexing parameters from YAML (17.5).}

\item{splice_junction_params}{(named list): Splice junction parameters from YAML (17.6)}

\item{log_file}{(str) The path of the log file to record events and errors.}
}
\value{
None The function logs progress and errors but returns no value.
}
\description{
prepare_STAR_genome
Preparing the STAR genome index directory using full genome indexing and splice junction parameters.
Checks whether a valid STAR genome index already exists in the specified
genome directory. If all required index files are present, it skips index generation.
Otherwise, it generates a new STAR genome index using the provided FASTA, genome indexing
parameters, and splice junction parameters.
}
\details{
Error codes:
ERROR CODE 11: STAR genome generation failed.
}
