% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_10x.R
\name{QC_10x}
\alias{QC_10x}
\title{QC_10x}
\usage{
QC_10x(output_folder, log_file)
}
\description{
Perform comprehensive QC on 10x Genomics data:
\itemize{
\item Load raw & filtered matrices
\item Compute QC metrics
\item Visualize QC (violin & bar plots)
\item Filter poor-quality cells
\item Detect potential doublets using Scrublet
}
}
