% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_log_message.R
\name{write_log}
\alias{write_log}
\title{add_log_message}
\usage{
write_log(message, log_file)
}
\arguments{
\item{message}{The actual commit message before push}

\item{log_file}{The path of the log file}
}
\description{
Log a message with a time stamp.
}
\keyword{internal}
