% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_start_genome.R
\name{prepare_STAR_genome}
\alias{prepare_STAR_genome}
\title{prepare_STAR_genome}
\usage{
prepare_STAR_genome(
  genome_dir,
  fasta_file,
  genome_index_params,
  splice_junction_params,
  log_file
)
}
\arguments{
\item{genome_dir}{The system-based absolute path of the directory where the STAR
genome index will be stored.}

\item{fasta_file}{The absolute path to the reference genome FASTA file.}

\item{log_file}{The path of the log file to record events and errors.}

\item{gtf_file}{The absolute path to the GTF annotation file.}

\item{read_length}{Integer, the sequencing read length. STAR will use (read_length - 1)
as the sjdbOverhang value.}
}
\value{
None The function logs progress and errors but returns no value.
}
\description{
Preparing the STAR genome index directory.
}
\details{
Checks whether a valid STAR genome index already exists in the specified
genome directory. If all required index files are present, the function
logs this and skips index generation. If not, it generates a new STAR
genome index using the provided FASTA and GTF annotation files.

Error codes:
ERROR CODE 8: STAR genome generation failed.

Behavior:
\itemize{
\item If genome_dir exists and contains valid STAR index files (Genome, SA, and SAindex), the generation
step is
skipped.
\item If the index is missing, a new one is created with STAR.
}
}
