% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_start_genome.R
\name{prepare_STAR_genome}
\alias{prepare_STAR_genome}
\title{Preparing the STAR genome index directory.}
\usage{
prepare_STAR_genome(genome_dir, fasta_file, gtf_file, read_length, log_file)
}
\arguments{
\item{genome_dir}{The system-based absolute path of the directory where the STAR
genome index will be stored.}

\item{fasta_file}{The absolute path to the reference genome FASTA file.}

\item{gtf_file}{The absolute path to the GTF annotation file.}

\item{read_length}{Integer, the sequencing read length. STAR will use (read_length - 1)
as the sjdbOverhang value.}

\item{log_file}{The path of the log file to record events and errors.}
}
\description{
Checks whether a valid STAR genome index already exists in the specified
genome directory. If all required index files are present, the function
logs this and skips index generation. If not, it generates a new STAR
genome index using the provided FASTA and GTF annotation files.
}
\details{
Error codes:
ERROR CODE 8: STAR genome generation failed.

Behavior:
\itemize{
\item If genome_dir exists and contains valid STAR index files (Genome, SA, and SAindex), the generation step is
skipped.
\item If the index is missing, a new one is created with STAR.
}

@return None The function logs progress and errors but returns no value.
}
